Fix non-returned value
If a state variable was not present, or had a bad/NaN value, an exception would be raised. In this case, the `_previous_data` attribute would not be updated, and the error would persist at future steps. This makes sure previous data is update AND make sure the input data is returned for other steps in the brokkr chain.